Move resource characterizations plot tests into resource metrics file
